bestproject2/microservices/product-catalog/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytest==8.3.2
testcontainers[postgresql,redis]==4.7.2
redis==5.0.1
elasticsearch==8.11.1
python-json-logger==2.0.7
prometheus-client==0.19.0
aiofiles==23.2.1
python-dotenv==1.0.0
alembic==1.13.1
